uqo-requests/backend/uqo_req/demandes/urls.py
from rest_framework.routers import DefaultRouter
from .views import DemandeViewSet

router = DefaultRouter()
router.register(r'demandes', DemandeViewSet, basename='demande')

urlpatterns = router.urls
