uqo-requests/backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
sqlparse==0.5.5
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
